backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cinetag-backend"
version = "0.1.0"
description = "CineTag distributed video intelligence backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "celery>=5.4",
  "redis>=5.0",
  "prometheus-client>=0.20",
  "httpx>=0.27",
  "openai>=1.30",
  "google-cloud-storage>=2.18",
  "google-cloud-pubsub>=2.25",
  "google-auth>=2.30",
  "google-auth-httplib2>=0.2",
  "google-cloud-secret-manager>=2.20",
  "pgvector>=0.3.6"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23"
]
[tool.ruff]
line-length = 100
[tool.pytest.ini_options]
pythonpath=["app"]
